refactor: moved StencilStart class to a separate file
